==> s5_summary.ipynb <==
import pandas as pd
import numpy as np
import os, pdb, sys
import matplotlib.pyplot as plt
pd.set_option('display.max_columns', None)

df = pd.read_csv('combined_dataset_with_vision.csv', sep = '~')
df.head()

for col in list(df.columns[-10:]): 
    sub = df[[col, 'desc']].groupby([col]).count().reset_index()
    sub = sub[sub[col] == 1.0].reset_index(drop = True)
    n_col = sub['desc'][0]
    print(f'{col}: {n_col} out of {df.shape[0]}')
    print(f'Percent: {n_col/df.shape[0]*100:.2f}%')
    print('---')
